// src/app/_restClientes/model/direccion.ts
/**
 * Servicio Clientes Kallsonys
 * Servicio para la gestión de clientes en  OMS Kallsonys
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Direccion { 
    iddireccion?: number;
    idcliente?: number;
    direccion?: string;
    pais?: string;
    ciudad?: string;
    tipodireccion?: string;
    estado?: string;
}